Quan Ly Khach Hang: chua xong phan regex phone, chua set confirm box cho delete button

+++ LTWWWJAVA_DHKTPM13A_BAITAPLON_NHOM24/src/main/java/fit/se/main/dao/verifyaccount/impl/VerifyAccountDAOImpl.java
@@ -24,7 +24,7 @@ public class VerifyAccountDAOImpl implements VerifyAccountDAO{
 	}
 
 	@Override
-	public Optional<VerifyAccount> findById(Long id) {
+	public Optional<VerifyAccount> findById(int id) {
 		return verifyAccountRepository.findById(id);
 	}
