+++ LTWWWJAVA_DHKTPM13A_BAITAPLON_NHOM24/src/main/java/fit/se/main/model/VerifyAccount.java
@@ -12,16 +12,13 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
-import javax.persistence.Lob;
 import javax.persistence.ManyToOne;
-
-import org.hibernate.annotations.Type;
 
 @Entity
 public class VerifyAccount {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Long id;
+	private int id;
 
 	@Column(name = "token", unique = true, length = 512)
 	private String token;
@@ -45,11 +42,11 @@ public class VerifyAccount {
 	public VerifyAccount() {
 	}
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
