+++ LTWWWJAVA_DHKTPM13A_BAITAPLON_NHOM24/src/main/java/fit/se/main/repository/VerifyAccountRepository.java
@@ -3,10 +3,9 @@ package fit.se.main.repository;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
 import fit.se.main.model.VerifyAccount;
 
-public interface VerifyAccountRepository extends JpaRepository<VerifyAccount, Long>{
+public interface VerifyAccountRepository extends JpaRepository<VerifyAccount, Integer>{
 	Optional<VerifyAccount> findByToken(String token);
 }
